+++ Modules/SWMB/LocalMachine-Security.psm1
@@ -1101,6 +1101,32 @@ Function TweakDisableBitlocker { # RESINFO
 	}
 }
 
+# View
+Function TweakViewBitlocker { # RESINFO
+	Write-Output "Viewing Bitlocker on all fixed drives (XtsAes256 Recommanded)..."
+	$ListVolume = Get-volume | Where-Object { $_.DriveType -eq "Fixed" }
+	ForEach ($Volume in $ListVolume) {
+		If (!($Volume.DriveLetter)) { Continue }
+
+		$Letter = $Volume.DriveLetter
+		$LetterColon = $Letter + ":"
+
+		# Test if partition is already encrypted (like for C:)
+		If ((Get-BitLockerVolume $Letter).ProtectionStatus -eq "On") {
+			# Get encryption method
+			$DriveEMethod = (Get-BitLockerVolume $Letter).EncryptionMethod
+			If ($DriveEMethod -eq "XtsAes256") {
+				Write-Output " Bitlocker on drive $Letter is OK (XtsAes256)"
+			} Else {
+				Write-Output " Bitlocker on drive $Letter is encrypted with $DriveEMethod (Please re-encrypt with XtsAes256)"
+			}
+			Continue
+		}
+
+		Write-Output " Bitlocker on drive $Letter is OFF (Please encrypt with XtsAes256)"
+	}
+}
+
 ################################################################
 
 # Suspend or Resume Bitlocker